+++ lib/luro_commands/src/functions/mod.rs
@@ -1,3 +1,4 @@
 pub mod saucenao;
 pub mod sendquote;
 pub mod sort_user_roles;
+pub mod userinfo;
