+++ lib/luro_events/src/events/member_joined.rs
@@ -7,7 +7,7 @@ pub async fn member_joined(ctx: &Context, user_data: &Data, accent_colour: [u8; 
     if let Some(alert_channel) = discod_event_log_channel_defined(&member.guild_id, user_data, ctx).await {
         let guild = member.guild_id.to_guild_cached(ctx);
 
-        let mut embed = event_embed(guild_accent_colour(accent_colour, guild), Some(&member.user), None).await;
+        let mut embed = event_embed(guild_accent_colour(accent_colour, guild), None, Some(&member.user)).await;
         embed.title("Member Joined").description(format!(
             "The user {} ({}) just joined the server!",
             member.display_name(),
